+++ src/app/(main)/page.tsx
@@ -1,10 +1,7 @@
 'use client';
-
 import { Box } from '@mui/material';
-import gsap from 'gsap';
-import { ScrollTrigger } from 'gsap/ScrollTrigger';
-import { useEffect, useRef } from 'react';
 
+import ScrollingFadeOut from '@/components/animations/ScrollingFadeOut';
 import { AppContentWrapper } from '@/components/common/Global.style';
 import About from '@/features/home/about/About';
 import Hero from '@/features/home/hero/Hero';
@@ -15,72 +12,18 @@ import Stats from '@/features/home/stats/Stats';
 import Support from '@/features/home/support/Support';
 import Testimonials from '@/features/home/testimonials/Testimonials';
 
-gsap.registerPlugin(ScrollTrigger);
-
 export default function Home() {
-  const contentWrapperRef = useRef(null);
-
-  useEffect(() => {
-    const tl = gsap.timeline({
-      scrollTrigger: {
-        trigger: '.scrolling-section',
-        start: 'top  center',
-        end: '+=30%',
-        scrub: true,
-        pin: true,
-        pinSpacing: false,
-        onUpdate: (self) => {
-          const opacity = 1 - self.progress;
-
-          gsap.set('.scrolling-section', { opacity: opacity });
-        }
-      },
-    });
-
-    tl.to(
-      '.scrolling-section',
-      {
-        opacity: 0,
-        duration: 0.5,
-      },
-      '<'
-    );
-    // .to(
-    //   contentWrapperRef.current,
-    //   {
-    //     scrollTrigger: {
-    //       trigger: ".scrolling-section",
-    //       start: "top center", 
-    //       end: "bottom bottom",
-
-    //       onEnter: () => {
-    //         window.scrollTo({
-    //           top: window.scrollY + window.innerHeight * 0.95,
-    //           behavior: "smooth"
-    //         });
-    //       },
-
-    //     },
-    //   },
-    //   ">"
-    // )
-
-    return () => {
-      ScrollTrigger.getAll().forEach((trigger) => trigger.kill());
-    };
-  }, []);
-
   return (
-    <Box ref={contentWrapperRef}>
+    <>
       <Hero />
       <Stats />
       <AppContentWrapper>
         <About />
       </AppContentWrapper>
 
-      <Box className="scrolling-section">
+      <ScrollingFadeOut>
         <Process />
-      </Box>
+      </ScrollingFadeOut>
       <Box
         sx={{
           backgroundColor: '#fffbfb',
@@ -97,6 +40,6 @@ export default function Home() {
           <Testimonials />
         </AppContentWrapper>
       </Box>
-    </Box>
+    </>
   );
 }
